+++ client/front/lib/map.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:developer';
 import 'package:flutter/material.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
-import 'package:front/errand/errand/request.dart';
+import 'package:front/errand/request.dart';
 import 'package:geolocator/geolocator.dart';
